fix(windows): support legacy SDK GUID parsing
Parse normalized UUID strings through IIDFromString so the operations backend builds against the supported Windows SDK.

Refs: butane_flutter-wisp-6c8.7

+++ packages/butane_windows/windows/test/butane_gatt_operations_winrt_test.cpp
@@ -127,7 +127,8 @@ TEST(WindowsGattOperationsBackend, MapsEveryErrorRow) {
   for (const auto& [status, code] : rows) {
     native->status = status;
     native->protocol = status == GattOperationStatus::kProtocolError
-                           ? std::optional<uint8_t>(0x0e)
+                           ? std::optional<uint8_t>(
+                                 static_cast<uint8_t>(0x0e))
                            : std::nullopt;
     backend.WriteDescriptor(1, "s", "c", "d", {}, [&](auto error) {
       ASSERT_TRUE(error);
